+++ CHANGELOG.md
@@ -57,6 +57,7 @@ Changes to error messages, help message and alike would **not** be considered br
 - **BREAKING**: `Compiler::compile` no longer returns an `Artifacts` struct, it is instead
     constructed beforehand and is now required in compilation steps instead of generic `PathBuf`s
 - **BREAKING**: Compilation steps are now `Step` associated functions `step_done` and `sub_step_done`
+- **BREAKING**: `<=>` now has the same precedence as `==`, `!=`, `<`, `<=`, `>`, `>=` operators
 
 ### Removed